Move state.js to src/main
