+++ src/main/index.js
@@ -13,7 +13,7 @@ const { getConfigPath } = require('../application-constants')
 const logins = require('./logins')
 const ipc = require('./ipc')
 const menu = require('./menu')
-const State = require('../renderer/lib/state')
+const State = require('./state')
 const windows = require('./windows')
 const logHandler = require('./developerTools/logHandler')
 const log = require('../logger').getLogger('main/index')
